Core implementation from Hack Upstate XIII (#1)
This commit adds a react-native client that will monitor current location and heading to determine
which side of the street you are parked on. It assumes parking rule data, and will count down until
the time to switch sides of the street.

Not yet implemented are Sunday rules and the changed display when parking and over the time limit. 

Buggy is the Navigator.getHeading, which only works when the phone is completely flat.